cocoa: fix some pointer casts to be 32bit safe
credits: wm4

+++ DOCS/client_api_examples/cocoabasic.m
@@ -82,7 +82,7 @@ static void wakeup(void *);
         }
 
 #if EMBED_VIEW
-        uintptr_t wid = (uintptr_t)self->w;
+        int64_t wid = (intptr_t) self->w;
         check_error(mpv_set_option(mpv, "wid", MPV_FORMAT_INT64, &wid));
 #endif
